Split hardware_interface.h into... - hardware_interface.h, which contains the actual interface to the hardware - hardware_types.h, a light weight class for pin related enums & names
